fix: c_char signedness varies i8(x86_64)/u8(aarch64) — unnecessary_cast on android
bore-android-tun's ifreq name pack/unpack cast bytes to/from c_char.
c_char is i8 on x86_64-linux-android but u8 on aarch64-linux-android,
so a direct `u8 as c_char`/`c_char as u8` is a real cast on one arch
and a same-type no-op on the other — clippy::unnecessary_cast (denied
via -D warnings) only fired for aarch64-linux-android, passing
x86_64-linux-android and every local gate (dev host is x86_64 Linux).
Route both casts through i32 (always a genuine widening cast on either
arch) so the cast is never flagged as redundant on either target.

+++ crates/bore-android-tun/src/lib.rs
@@ -44,7 +44,8 @@ pub fn create(name: &str) -> Result<(tun_rs::AsyncDevice, String)> {
 
     let mut ifr: nix::libc::ifreq = unsafe { std::mem::zeroed() };
     for (dst, src) in ifr.ifr_name.iter_mut().zip(name.as_bytes()) {
-        *dst = *src as std::os::raw::c_char;
+        // Same cross-arch c_char signedness note as `actual_name` below.
+        *dst = *src as i32 as std::os::raw::c_char;
     }
     ifr.ifr_ifru.ifru_flags = (nix::libc::IFF_TUN | nix::libc::IFF_NO_PI) as std::os::raw::c_short;
 
@@ -60,9 +61,12 @@ pub fn create(name: &str) -> Result<(tun_rs::AsyncDevice, String)> {
     let actual_name = {
         let raw = &ifr.ifr_name;
         let end = raw.iter().position(|&c| c == 0).unwrap_or(raw.len());
+        // `c_char` is `i8` on x86_64 but `u8` on aarch64 — route through `i32` so
+        // neither cast below is ever a same-type no-op on either arch (clippy
+        // ::unnecessary_cast fires per-target otherwise; interface names are ASCII).
         raw[..end]
             .iter()
-            .map(|&c| c as u8 as char)
+            .map(|&c| c as i32 as u8 as char)
             .collect::<String>()
     };
